OpenAPI not Swagger since 2016
This was discussed as a v1.0 feature but v1.0 has sailed, so im sending this PR to help get it done for v2.0.

Swagger is a) trademark of SmartBear, b) used almost exclusively by old crap unmaintained projects which do not support OpenAPI v3.0.

Your project is neither, but its gonna get confused for the latter because of the name. Please update to call it OpenAPI.