Fix pathMatcher to prioritize exact matches
Previously if multiple routes were defined like so:
/foo/exact_bar
/foo/:some_bar_variable

and the openapi definition happened to have the /foo/:some_bar_variable
defined first, the pathMatcher would consider this the correct path on
which to validate the request.

This change now does the pathMatcher in 2 phases: an exact match first,
followed by the previous logic of looking for a variable starting with
':'. This way the exact match is prioritized over an inexact match.